refactor(repository): move interface to interfaces package
ClassStaffRepository 인터페이스를 interfaces 패키지로 이동
